+++ sortedcontainers_pydantic/__init__.py
@@ -1,8 +1,11 @@
+from dataclasses import dataclass
+from functools import partial
 import importlib.metadata
 from typing import (
     TYPE_CHECKING,
+    Annotated,
     Any,
-    Generic,
+    Callable,
     Hashable,
     Iterable,
     Mapping,
@@ -10,6 +13,7 @@ from typing import (
     Tuple,
     TypeVar,
     get_args,
+    get_origin,
 )
 
 from pydantic import (
@@ -18,34 +22,71 @@ from pydantic import (
 from pydantic_core import core_schema
 import sortedcontainers
 
+if TYPE_CHECKING:
+    from _typeshed import SupportsRichComparison
+
+
 __version__ = importlib.metadata.version("sortedcontainers-pydantic")
+
+__all__ = [
+    "SortedDict",
+    "SortedList",
+    "SortedSet",
+    "SortedKeyList",
+    "SortedDictPydanticAnnotation",
+    "SortedListPydanticAnnotation",
+    "SortedSetPydanticAnnotation",
+    "AnnotatedSortedDict",
+    "AnnotatedSortedList",
+    "AnnotatedSortedSet",
+    "Key",
+    "UnsupportedSourceTypeError",
+]
 
 _KT = TypeVar("_KT", bound=Hashable)  # Key type.
 _VT = TypeVar("_VT")  # Value type.
 _T = TypeVar("_T")
+_OrderableT = TypeVar("_OrderableT", bound="SupportsRichComparison")
 _HashableT = TypeVar("_HashableT", bound=Hashable)
 
-# sortedcontainers is not type annotated directed so we need this hack to declare the parent
-# classes as generics for Python 3.8.
-# https://mypy.readthedocs.io/en/stable/runtime_troubles.html#using-classes-that-are-generic-in-stubs-but-not-at-runtime
-if TYPE_CHECKING:
 
-    class _SortedDict(sortedcontainers.SortedDict[_KT, _VT]): ...
-
-    class _SortedList(sortedcontainers.SortedList[_T]): ...
-
-    class _SortedSet(sortedcontainers.SortedSet[_HashableT]): ...
-
-else:
-
-    class _SortedDict(Generic[_KT, _VT], sortedcontainers.SortedDict): ...
-
-    class _SortedList(Generic[_T], sortedcontainers.SortedList): ...
-
-    class _SortedSet(Generic[_HashableT], sortedcontainers.SortedSet): ...
+class _UnsupportedSourceTypeError(Exception):
+    def __init__(self, parsed: Any):
+        self.parsed = parsed
 
 
-class SortedDict(_SortedDict[_KT, _VT]):
+class UnsupportedSourceTypeError(TypeError):
+    pass
+
+
+def _get_constructor(tp: Any) -> Any:
+    """Get the relevant class constructor for the given type annotation, e.g., SortedList from
+    SortedList[int] or appropriate subclass.
+    """
+    # Parse the annotation to get the relevant source type, e.g., SortedList form SortedList[int]
+    origin = get_origin(tp)
+    if origin is None:
+        parsed = tp
+    else:
+        parsed = origin
+
+    bases = set((parsed,) + getattr(parsed, "__bases__", tuple()))
+    if bases & {
+        # Subclasses of sortedcontainers_pydantic
+        SortedList,
+        SortedDict,
+        SortedSet,
+        # Subclasses of sortedcontainers
+        sortedcontainers.SortedList,
+        sortedcontainers.SortedDict,
+        sortedcontainers.SortedSet,
+    }:
+        return parsed
+    else:
+        raise _UnsupportedSourceTypeError(parsed)
+
+
+class SortedDictPydanticAnnotation:
     @classmethod
     def __get_pydantic_core_schema__(
         cls, source_type: Any, handler: GetCoreSchemaHandler
@@ -61,14 +102,25 @@ class SortedDict(_SortedDict[_KT, _VT]):
             - If it's an iterable, pass to SortedList constructor
         - Serialization: Convert to a list
         """
+        if cls is SortedDictPydanticAnnotation:
+            # Used as annotation, i.e., Annotated[..., SortedDictPydanticAnnotation]
+            try:
+                cls = _get_constructor(source_type)
+            except _UnsupportedSourceTypeError as e:
+                msg = (
+                    "Expected subclass of sortedcontainers.SortedDict, "
+                    f"got '{e.parsed}' parsed from annotation '{source_type}'."
+                )
+                raise UnsupportedSourceTypeError(msg) from e
+
         # Schema for when the input is already an instance of this class
         instance_schema = core_schema.is_instance_schema(cls)
 
         # Get schema for Iterable type based on source type has arguments
         args = get_args(source_type)
         if args:
-            mapping_t_schema = handler.generate_schema(Mapping[args[0], args[1]])  # type: ignore
-            iterable_of_pairs_t_schema = handler.generate_schema(Iterable[Tuple[args[0], args[1]]])  # type: ignore
+            mapping_t_schema = handler.generate_schema(Mapping[args[0], args[1]])  # type: ignore[valid-type]
+            iterable_of_pairs_t_schema = handler.generate_schema(Iterable[Tuple[args[0], args[1]]])  # type: ignore[valid-type]
         else:
             mapping_t_schema = handler.generate_schema(Mapping)
             iterable_of_pairs_t_schema = handler.generate_schema(Iterable[Tuple[Any, Any]])
@@ -83,16 +135,18 @@ class SortedDict(_SortedDict[_KT, _VT]):
             function=cls, schema=iterable_of_pairs_t_schema
         )
 
-        # Union of the two schemas
-        python_schema = core_schema.union_schema(
-            [
-                instance_schema,
-                from_mapping_schema,
-                from_iterable_of_pairs_schema,
-            ]
-        )
+        # Union the schemas
+        # Only include instance_schema if there are no type arguments
+        # Otherwise an existing instance with wrong argument types won't be coerced
+        if args:
+            python_schema = core_schema.union_schema(
+                [from_mapping_schema, from_iterable_of_pairs_schema]
+            )
+        else:
+            python_schema = core_schema.union_schema(
+                [instance_schema, from_mapping_schema, from_iterable_of_pairs_schema]
+            )
 
-        # Serializer that converts an instance to a dict
         as_dict_serializer = core_schema.plain_serializer_function_ser_schema(dict)
 
         return core_schema.json_or_python_schema(
@@ -102,7 +156,16 @@ class SortedDict(_SortedDict[_KT, _VT]):
         )
 
 
-class SortedList(_SortedList[_T]):
+class SortedDict(sortedcontainers.SortedDict[_KT, _VT], SortedDictPydanticAnnotation):
+    pass
+
+
+AnnotatedSortedDict = Annotated[
+    sortedcontainers.SortedDict[_KT, _VT], SortedDictPydanticAnnotation
+]
+
+
+class SortedListPydanticAnnotation:
     @classmethod
     def __get_pydantic_core_schema__(
         cls, source_type: Any, handler: GetCoreSchemaHandler
@@ -118,13 +181,24 @@ class SortedList(_SortedList[_T]):
             - If it's an iterable, pass to SortedList constructor
         - Serialization: Convert to a list
         """
+        if cls is SortedListPydanticAnnotation:
+            # Used as annotation, e.g., Annotated[..., SortedListPydanticAnnotation]
+            try:
+                cls = _get_constructor(source_type)
+            except _UnsupportedSourceTypeError as e:
+                msg = (
+                    "Expected subclass of sortedcontainers.SortedList, "
+                    f"got '{e.parsed}' parsed from annotation '{source_type}'."
+                )
+                raise UnsupportedSourceTypeError(msg) from e
+
         # Schema for when the input is already an instance of this class
         instance_schema = core_schema.is_instance_schema(cls)
 
         # Get schema for Iterable type based on source type has arguments
         args = get_args(source_type)
         if args:
-            iterable_t_schema = handler.generate_schema(Iterable[args[0]])  # type: ignore
+            iterable_t_schema = handler.generate_schema(Iterable[args[0]])  # type: ignore[valid-type]
         else:
             iterable_t_schema = handler.generate_schema(Iterable)
 
@@ -134,12 +208,13 @@ class SortedList(_SortedList[_T]):
         )
 
         # Union of the two schemas
-        python_schema = core_schema.union_schema(
-            [
-                instance_schema,
-                from_iterable_schema,
-            ]
-        )
+        # Only include instance_schema if there are no type arguments
+        # Otherwise an existing instance with wrong argument types won't be coerced
+        python_schema: core_schema.CoreSchema
+        if args:
+            python_schema = from_iterable_schema
+        else:
+            python_schema = core_schema.union_schema([instance_schema, from_iterable_schema])
 
         # Serializer that converts an instance to a list
         as_list_serializer = core_schema.plain_serializer_function_ser_schema(list)
@@ -151,7 +226,22 @@ class SortedList(_SortedList[_T]):
         )
 
 
-class SortedSet(_SortedSet[_HashableT]):
+class SortedList(sortedcontainers.SortedList[_T], SortedListPydanticAnnotation):
+    pass
+
+
+class SortedKeyList(sortedcontainers.SortedKeyList[_T, _OrderableT], SortedList[_T]):
+    pass
+
+
+AnnotatedSortedList = Annotated[sortedcontainers.SortedList[_T], SortedListPydanticAnnotation]  # type: ignore[misc]
+
+# Don't define AnnotatedSortedKeyList
+# We generally don't expect people to directly use SortedKeyList in a field definition
+# They should use SortedList instead
+
+
+class SortedSetPydanticAnnotation:
     @classmethod
     def __get_pydantic_core_schema__(
         cls, source_type: Any, handler: GetCoreSchemaHandler
@@ -168,14 +258,25 @@ class SortedSet(_SortedSet[_HashableT]):
             - If it's an iterable, pass to SortedSet constructor
         - Serialization: Convert to a list
         """
+        if cls is SortedSetPydanticAnnotation:
+            # Used as annotation, i.e., Annotated[..., SortedSetPydanticAnnotation]
+            try:
+                cls = _get_constructor(source_type)
+            except _UnsupportedSourceTypeError as e:
+                msg = (
+                    "Expected subclass of sortedcontainers.SortedSet, "
+                    f"got '{e.parsed}' parsed from annotation '{source_type}'."
+                )
+                raise UnsupportedSourceTypeError(msg) from e
+
         # Schema for when the input is already an instance of this class
         instance_schema = core_schema.is_instance_schema(cls)
 
         # Get schema for Iterable type based on source type has arguments
         args = get_args(source_type)
         if args:
-            set_t_schema = handler.generate_schema(Set[args[0]])  # type: ignore
-            iterable_t_schema = handler.generate_schema(Iterable[args[0]])  # type: ignore
+            set_t_schema = handler.generate_schema(Set[args[0]])  # type: ignore[valid-type]
+            iterable_t_schema = handler.generate_schema(Iterable[args[0]])  # type: ignore[valid-type]
         else:
             set_t_schema = handler.generate_schema(Set)
             iterable_t_schema = handler.generate_schema(Iterable)
@@ -190,14 +291,15 @@ class SortedSet(_SortedSet[_HashableT]):
             function=cls, schema=iterable_t_schema
         )
 
-        # Union of the two schemas
-        python_schema = core_schema.union_schema(
-            [
-                instance_schema,
-                from_set_schema,
-                from_iterable_schema,
-            ]
-        )
+        # Union of the schemas
+        # Only include instance_schema if there are no type arguments
+        # Otherwise an existing instance with wrong argument types won't be coerced
+        if args:
+            python_schema = core_schema.union_schema([from_set_schema, from_iterable_schema])
+        else:
+            python_schema = core_schema.union_schema(
+                [instance_schema, from_set_schema, from_iterable_schema]
+            )
 
         # Serializer that converts an instance to a list
         as_list_serializer = core_schema.plain_serializer_function_ser_schema(list)
@@ -206,4 +308,36 @@ class SortedSet(_SortedSet[_HashableT]):
             json_schema=from_set_schema,
             python_schema=python_schema,
             serialization=as_list_serializer,
+        )
+
+
+class SortedSet(sortedcontainers.SortedSet[_HashableT], SortedSetPydanticAnnotation):
+    pass
+
+
+AnnotatedSortedSet = Annotated[sortedcontainers.SortedSet[_HashableT], SortedSetPydanticAnnotation]
+
+
+@dataclass(frozen=True)
+class Key:
+    key: Callable[[Any], "SupportsRichComparison"]
+
+    def __get_pydantic_core_schema__(
+        self, source_type: Any, handler: GetCoreSchemaHandler
+    ) -> core_schema.CoreSchema:
+        try:
+            constructor = _get_constructor(source_type)
+        except _UnsupportedSourceTypeError as e:
+            msg = (
+                "Expected subclass of a sortedcontainers or sortedcontainers_pydantic class, "
+                f"got '{e.parsed}' parsed from annotation '{source_type}'."
+            )
+            raise UnsupportedSourceTypeError(msg) from e
+        # sortedcontainers.SortedList has magic behavior where the SortedList constructor returns
+        # SortedKeyList if given a key. We match that behavior for our SortedList.
+        if constructor is SortedList:
+            constructor = SortedKeyList
+        return core_schema.no_info_after_validator_function(
+            function=partial(constructor, key=self.key),
+            schema=handler(source_type),
         )